requirements.txt
# Core Flask dependencies
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# HTTP and networking
requests==2.31.0
urllib3==2.1.0

# Environment and configuration
python-dotenv==1.0.0

# Database and storage
supabase==2.22.1
postgrest==2.22.1

# LLM and AI
groq==0.4.2
numpy==1.24.3
scikit-learn==1.3.2

# Emotion Detection
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99
accelerate==0.25.0

# Authentication and security
pyjwt[crypto]==2.10.1
cryptography==45.0.6

# Monitoring and logging
sentry-sdk[flask]==1.40.0

# Data processing
pydantic==2.11.7

# Optional: Redis for caching
redis==5.0.1

# Optional: Metrics and monitoring
prometheus-client==0.19.0

# Optional: Streamlit for admin dashboard
streamlit==1.28.1

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Production server
gunicorn==21.2.0